Added option for no platform and module parameter to select it.
This enables platforms that can work with base Vulkan (for example for compute calculations only) to not need to worry about any other platform requirements. This will be useful if we are compiling for a platform not contemplated on the Jai compiler or the Vulkan Headers, or if you simply don't need any of the platform specific stuff and you don't want to worry about having those extra types and import the platform code for it to compile.

To select this I also added a module parameter, which for now will default to a best guess if not provided. With the current Jai compiler version I'm using you're not able to select it properly since the enum with the Vulkan platforms is not declared at the time of importing the module cause it's on the module itself.