Fixing device-level/instance-extension commands.
This moves device-level commands that are included in instance extensions to be loaded with the rest of the instance-level commands.

Seems to be bit confusing. But the spec indicates that even though the commands might be device level, if they are added through a instance-level extension they should be loaded with vkGetInstanceProcAddr and not with vkGetDeviceProcAddr.

See: https://github.com/zeux/volk/issues/8 for more details.